+++ tests/example_plugin.py
@@ -1,25 +1,24 @@
 import dataclasses
 import logging
-import re
 import typing
 
-from gather_vision.plugin import data as plugin_data, entry as plugin_entry
+from gather_vision.plugin import data, entry
 
 logger = logging.getLogger(__name__)
 
 logging.getLogger("example_plugin").setLevel("INFO")
 
 
-class ExamplePlugin(plugin_entry.Entry):
+class ExamplePlugin(entry.Entry):
     plugin_name = "example-plugin"
     plugin_value = "example_plugin.ExamplePlugin"
     _data_source_1 = "example-data-source-1"
     _data_source_2 = "example-data-source-2"
     _data_source_names = [_data_source_1, _data_source_2]
 
-    def update(self, args: plugin_entry.UpdateArgs) -> plugin_entry.UpdateResult:
+    def update(self, args: entry.UpdateArgs) -> entry.UpdateResult:
         logger.info(f"Running update for plugin {self.plugin_name}.")
-        return plugin_entry.UpdateResult(
+        return entry.UpdateResult(
             web_data=[
                 ExamplePluginWebData(
                     plugin_name=self.plugin_name,
@@ -34,21 +33,21 @@ class ExamplePlugin(plugin_entry.Entry):
             ],
         )
 
-    def list(self, args: plugin_entry.ListArgs) -> plugin_entry.ListResult:
+    def list(self, args: entry.ListArgs) -> entry.ListResult:
         logger.info(f"Running list for plugin {self.plugin_name}.")
-        return plugin_entry.ListResult(
+        return entry.ListResult(
             items={self.plugin_name: self._data_source_names},
         )
 
 
-class ExamplePluginWebData(plugin_data.WebData):
+class ExamplePluginWebData(data.WebData):
     def initial_urls(self) -> typing.Iterable[str]:
         return ["https://example.com/"]
 
     def parse_response(
-        self, data: plugin_data.WebDataAvailable
+        self, data: data.WebDataAvailable
     ) -> typing.Generator[
-        typing.Union[str, plugin_data.GatherDataItem], typing.Any, typing.Any
+        typing.Union[str, data.GatherDataItem], typing.Any, typing.Any
     ]:
         yield ExamplePluginDataItem(
             plugin_name=self.plugin_name,
@@ -65,15 +64,15 @@ class ExamplePluginWebData(plugin_data.WebData):
 
 
 @dataclasses.dataclass
-class ExamplePluginDataItem(plugin_data.GatherDataItem):
+class ExamplePluginDataItem(data.GatherDataItem):
     name: str
     value: int
 
 
-class ExamplePluginLocalData(plugin_data.LocalData):
+class ExamplePluginLocalData(data.LocalData):
     def load_resources(
         self,
-    ) -> typing.Generator[plugin_data.GatherDataItem, typing.Any, typing.Any]:
+    ) -> typing.Generator[data.GatherDataItem, typing.Any, typing.Any]:
         items = [
             ExamplePluginDataItem(
                 plugin_name=self.plugin_name,
@@ -91,8 +90,3 @@ class ExamplePluginLocalData(plugin_data.LocalData):
 
         for item in items:
             yield item
-
-
-def collapse_whitespace(value: str) -> str:
-    whitespace = re.compile(r"\s+")
-    return whitespace.sub(" ", value)